% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana-package.R
\docType{package}
\name{banana-package}
\alias{banana}
\alias{banana-package}
\title{banana: Translate everything into Banana size.}
\description{
Translate everything into bananas.
}
\author{
\strong{Maintainer}: Roel M. Hogervorst \email{hogervorst.rm@gmail.com} (\href{https://orcid.org/0000-0001-7509-0328}{ORCID})

}
\keyword{internal}
