% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{banana}
\alias{banana}
\alias{is_banana}
\alias{as_banana}
\title{`banana` vector}
\usage{
banana(x = double())

is_banana(x)

as_banana(x)
}
\arguments{
\item{x}{object to convert into banana}
}
\value{
An S3 vector of class `banana`.
}
\description{
This creates a double vector that represents banana so when it is
printed, it is converted and displayed as ğŸŒ.
However the underlying structure is still units vector.
}
\examples{
banana(c(0.25, 0.5, 0.75))
}
